implement saving a gig

+++ gighub/View_Models/GigViewModel.cs
@@ -8,7 +8,7 @@ namespace gighub.View_Models
         public string Venue { get; set; }
         public string Date { get; set; }
         public string Time { get; set; }
-        public int Genre { get; set; }
+        public byte Genre { get; set; }
         public IEnumerable Genres { get; set; }
     }
 }
